AuthService uses UserService
Cleaned up names and privacy

+++ backend/src/queries/follow_requests.sql
@@ -8,7 +8,7 @@ DELETE FROM follow_requests WHERE requester_id = :requesterId AND requestee_id =
 DELETE FROM follow_requests WHERE requester_id = :requesterId AND requestee_id = :requesteeId RETURNING *;
 
 /* @name Count */
-SELECT COUNT(*) FROM follow_requests WHERE requester_id = :requesterId AND requestee_id = :requesteeId;
+SELECT COUNT(*)::int FROM follow_requests WHERE requester_id = :requesterId AND requestee_id = :requesteeId;
 
 /* @name GetByRequesteeId */
 SELECT requester_id FROM follow_requests WHERE requestee_id = :requesteeId;
